+++ CHANGELOG.md
@@ -40,6 +40,9 @@
   [woodhamgh](https://github.com/woodhamgh)
   [696](https://github.com/realm/SwiftLint/issues/696)
 
+* Adds 'ConditionalReturnsOnNewLineRule' rule.  
+  [Rohan Dhaimade](https://github.com/HaloZero)
+
 ##### Bug Fixes
 
 * Fixed CustomRule Regex.